Refactor LatentCliqueLifting as a composed lifting map
